requirements.txt
opencv-python==4.8.1.78
opencv-contrib-python==4.8.1.78
numpy==1.24.3
Pillow==10.1.0
streamlit==1.28.0
streamlit-drawable-canvas==0.9.3
scipy==1.11.3
matplotlib==3.8.0
scikit-image==0.22.0

